Move package related code into separate translation unit `package`
